pyproject.toml
[project]
name = "cluster-alns"
version = "0.1.0"
description = "An Intelligent Clustering-Based ALNS: A Strong Baseline for Neural Combinatorial Optimization"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "absl-py==2.1.0",
    "alns==7.0.0",
    "astunparse==1.6.3",
    "certifi==2024.12.14",
    "charset-normalizer==3.4.1",
    "cloudpickle==3.1.1",
    "colorama==0.4.6",
    "contourpy==1.3.1",
    "cycler==0.12.1",
    "farama-notifications==0.0.4",
    "filelock==3.17.0",
    "flatbuffers==25.1.24",
    "fonttools==4.55.5",
    "fsspec==2024.12.0",
    "gast==0.6.0",
    "google-pasta==0.2.0",
    "grpcio==1.70.0",
    "gym==0.26.2",
    "gym-notices==0.0.8",
    "gymnasium==1.0.0",
    "h5py==3.12.1",
    "idna==3.10",
    "ipython>=8.37.0",
    "isort>=7.0.0",
    "jinja2==3.1.5",
    "joblib>=1.5.2",
    "keras==3.8.0",
    "kiwisolver==1.4.8",
    "kneed>=0.8.5",
    "libclang==18.1.1",
    "markdown==3.7",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "matplotlib==3.10.0",
    "mdurl==0.1.2",
    "ml-dtypes==0.4.1",
    "mpmath==1.3.0",
    "namex==0.0.8",
    "networkx==3.4.2",
    "numpy==1.26.4",
    "opt-einsum==3.4.0",
    "optree==0.14.0",
    "packaging==24.2",
    "pandas==2.2.3",
    "pillow==11.1.0",
    "protobuf==5.29.3",
    "pygame==2.6.1",
    "pygments==2.19.1",
    "pyparsing==3.2.1",
    "python-dateutil==2.9.0.post0",
    "pytz==2024.2",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "rich==13.9.4",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "shimmy==2.0.0",
    "six==1.17.0",
    "stable-baselines3==2.4.1",
    "sympy==1.13.1",
    "tensorboard==2.18.0",
    "tensorboard-data-server==0.7.2",
    "tensorflow==2.18.0",
    "tensorflow-io-gcs-filesystem==0.31.0",
    "termcolor==2.5.0",
    "torch==2.5.1",
    "tqdm==4.67.1",
    "typing-extensions==4.12.2",
    "tzdata==2025.1",
    "urllib3==2.3.0",
    "wandb>=0.23.0",
    "werkzeug==3.1.3",
    "wrapt==1.17.2",
]

[tool.setuptools.packages.find]
include = ["cluster_alns*"]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "isort>=7.0.0",
    "ruff>=0.14.7",
]

[tool.ruff]
select = ["I"]
